Move bucketfs api to separate package
