+++ pkg/extensionController/bfs/BucketFsApi.go
@@ -1,4 +1,4 @@
-package extensionController
+package bfs
 
 import (
 	"context"
