+++ pkg/extensionController/bfs/BucketFsApi_test.go
@@ -1,4 +1,4 @@
-package extensionController
+package bfs
 
 import (
 	"context"
